requirements.txt
playwright==1.41.2
pytest==7.4.3
Pillow==10.1.0
jinja2==3.1.2
